Fix MongoDB stored results rendering empty table in event history
MongoDB queries return null columns, causing the stored results table to
render with no headers or cell data. Use the existing JsonViewer component
(extracted into its own file) for MongoDB stored results instead.
